# contracts/nodes-v1.openapi.yaml
openapi: 3.1.0
info:
  title: RETINA node ingest
  version: 1.4.0
  description: |
    The RETINA server's HTTP API. The paths under `/v1/nodes` are the RETINA node
    ingest contract, generated from the server and versioned as a unit; everything
    else is internal to the map, the dashboard and the admin surfaces. The tower
    search this app was named for now lives in tower-finder-service, which every
    vhost reaches directly through nginx.

    ## Error taxonomy

    Every refusal under `/v1/nodes` carries `Error`, whose required key is `error`
    and whose optional `detail` names a field rather than describing it. The rest of
    this API answers in FastAPI's `{"detail": ...}`, which its own callers parse.

    | Status | Means |
    |---|---|
    | `400` | The body was refused. `invalid_config` for a configuration that failed validation, `invalid_contact` for contact details that failed it, `invalid_claim` for an address that failed it, `invalid_body` for a body that failed the schema |
    | `401` | The bearer token is bad, revoked or expired |
    | `403` | Registration refused, without saying why |
    | `409` | The frame names a `config_version` this server never issued, or the node already has an owner |
    | `413` | The body exceeded the cap for its path |
    | `429` | Rate limited |

    One refusal does not carry `Error`, and it is the only one: the `409` on
    `PUT /v1/nodes/claim` answers with `ClaimResponse` instead. Nothing the node can
...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/nodes/contact:
    put:
      tags:
      - contact
      summary: Report the owner's contact details.
      description: |-
        Whom to contact about this node. Sent when the details change locally, and not otherwise:
        nothing on the server asks for them, and no response marks them stale.

        A node with nothing to report need never call this. The `NodeContact` schema says what
        each field means and what omitting one does.

        What is stored is unverified and is used to reach the owner about their own node. It is
        not an account, and it grants nothing.
      operationId: putContact
      requestBody:
        description: The full contact document. Fields omitted are cleared.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeContact'
        required: true
      responses:
        '200':
          description: When the stored details last changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResponse'
        '400':
          description: The contact details failed validation, as `invalid_contact`. A body that is not
            JSON at all lands here too, since the remedy is the same. `detail` names the offending field
            and nothing else, so retina-gui can put the message next to the input it belongs to. Retrying
            unchanged will not help.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: never
          x-terminal: false
        '401':
          description: 'Token bad, revoked or expired. Surface it locally and leave the details unsent:
            they are sent again whenever they change, so nothing is lost by waiting for a credential.
            Do not re-register, for the reason detection and heartbeat give on their own 401s.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: never
          x-terminal: true
        '413':
          description: 'The body exceeded this path''s cap, which is checked from `Content-Length` before
            parsing. Per-field bounds in the schemas do not make this unreachable: `HeartbeatRequest.errors`
            alone permits 16 KiB of strings, `RegisterRequest.config` carries no bound at all, and JSON
            puts no length limit on a number''s text, so a schema-valid body can still exceed it.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: never
          x-terminal: false
        5XX:
          description: 'Server side, though a 502 or 504 is the edge''s own response rather than the origin''s,
            and carries whatever body that gateway sends instead of this `Error`. Abandon the request
            and retry, backing off with jitter if it persists: on detection and heartbeat that retry is
            the node''s next frame or beat rather than a resend of this one, while registration and configuration
            retry the identical body.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
          x-retry: backoff
          x-terminal: false
      security:
      - bearerAuth: []
      x-cadence: on local change only
      x-max-body-bytes: 2048
...
components:
  schemas:
...
    ContactResponse:
      properties:
        updated_at:
          type: string
          format: date-time
          title: Updated At
        claim_state:
          type: string
          enum:
          - unclaimed
          - pending
          - owned
          title: Claim State
        claim_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Claim Email
        claim_undeliverable:
          type: boolean
          title: Claim Undeliverable
      type: object
      required:
      - updated_at
      - claim_state
      - claim_email
      - claim_undeliverable
      title: ContactResponse
...
    ErrorBody:
      properties:
        error:
          type: string
          maxLength: 64
          title: Error
        detail:
          anyOf:
          - type: string
            maxLength: 512
          - type: 'null'
          title: Detail
      type: object
      required:
      - error
      title: ErrorBody
      description: |-
        The shape every node-API refusal wears. `error` is a stable slug;
        `detail`, present only when it applies, names the offending field or value
        rather than describing it. Registration's own refusals never carry
        `detail`, by design.
...
    NodeContact:
      type: object
      title: NodeContact
      description: |-
        Whom to contact about this node, as its owner gave them. Every field is optional
        and nullable, and an absent field means the same as a null one: the document is
        replaced wholesale, so a field left out is cleared.

        `country` is an ISO 3166-1 alpha-2 code and says where the phone number is,
        which a national number cannot say for itself: `07700900123` is dialable only
        by someone who already knows the answer. Two letters, case-insensitive on the
        wire and stored upper case.

        Necessary but not sufficient, as the configuration schema is: surrounding
        whitespace is stripped before a value is judged, so a bound here describes the
        trimmed form rather than the bytes sent.

        None of it is verified, and none of it identifies anyone to the server: it is
        carried so that a fault we can see and the owner cannot has somewhere to go.
      properties:
        first_name:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
        last_name:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
        email:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
        phone:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
        country:
          anyOf:
          - type: string
            maxLength: 2
            pattern: ^[A-Za-z]{2}$
          - type: 'null'
      required: []
      additionalProperties: false
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'The token minted by `POST /v1/nodes/register`, persisted at mode 0600 under `/data`.
        Sent as `Authorization: Bearer <token>`. There is no expiry: a token dies by being revoked and
        by nothing else.'
      scheme: bearer